Android: Log exception stack traces and show debug information button in settings
