+++ android/src/main/java/org/ergoplatform/android/App.kt
@@ -5,8 +5,14 @@ import androidx.appcompat.app.AppCompatDelegate
 import org.ergoplatform.NodeConnector
 import org.ergoplatform.appkit.NetworkType
 import org.ergoplatform.isErgoMainNet
+import org.ergoplatform.utils.LogUtils
 
 class App : Application() {
+
+    companion object {
+        var lastStackTrace: String? = null
+            private set
+    }
 
     override fun onCreate() {
         super.onCreate()
@@ -14,5 +20,7 @@ class App : Application() {
         val preferences = Preferences(applicationContext)
         AppCompatDelegate.setDefaultNightMode(preferences.dayNightMode)
         NodeConnector.getInstance().loadPreferenceValues(preferences)
+
+        LogUtils.stackTraceLogger = { lastStackTrace = it }
     }
 }
